Thema instellingen naar Theme.js, dark theme toegevoegd, ingekort
